Refactor binding name to silvercrestwifisocket.
