--- addons/binding/org.openhab.binding.silvercrest/ESH-INF/binding/binding.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="silvercrest"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
-        xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
-
-    <name>Silvercrest Binding</name>
-    <description>This is the binding for Silvercrest brand things.</description>
-    <author>Jaime Vaz</author>
-
-</binding:binding>
