--- addons/binding/org.openhab.binding.silvercrest/OSGI-INF/SilvercrestHandlerFactory.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2014 openHAB UG (haftungsbeschraenkt) and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.binding.silvercrest.internal.SilvercrestHandlerFactory">
-
-   <implementation class="org.openhab.binding.silvercrest.internal.SilvercrestHandlerFactory"/>
-
-   <service>
-      <provide interface="org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory"/>
-   </service>
-
-</scr:component>
