--- addons/binding/org.openhab.binding.silvercrest/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.openhab.binding</groupId>
-		<artifactId>pom</artifactId>
-		<version>2.0.0-SNAPSHOT</version>
-	</parent>
-
-	<properties>
-		<bundle.symbolicName>org.openhab.binding.silvercrest</bundle.symbolicName>
-		<bundle.namespace>org.openhab.binding.silvercrest</bundle.namespace>
-	</properties>
-
-	<groupId>org.openhab.binding</groupId>
-	<artifactId>org.openhab.binding.silvercrest</artifactId>
-	<version>2.0.0-SNAPSHOT</version>
-
-	<name>Silvercrest Binding</name>
-	<packaging>eclipse-plugin</packaging>
-
-</project>
