--- addons/binding/org.openhab.binding.silvercrest/src/main/java/org/openhab/binding/silvercrest/SilvercrestBindingConstants.java
@@ -1,73 +1,0 @@
-/**
- * Copyright (c) 2014 openHAB UG (haftungsbeschraenkt) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.openhab.binding.silvercrest;
-
-import java.util.Set;
-
-import org.eclipse.smarthome.core.thing.ThingTypeUID;
-
-import com.google.common.collect.ImmutableSet;
-
-/**
- * The {@link SilvercrestBindingConstants} class defines common constants, which are
- * used across the whole binding.
- *
- * @author Jaime Vaz - Initial contribution
- */
-public class SilvercrestBindingConstants {
-
-    /**
-     * The binding id.
-     */
-    public static final String BINDING_ID = "silvercrest";
-
-    /**
-     * List of all Thing Type UIDs.
-     */
-    public final static ThingTypeUID THING_TYPE_SOCKET_OUTLET = new ThingTypeUID(BINDING_ID, "wifiSocketOutlet");
-
-    /**
-     * List of all Channel ids
-     */
-    public final static String SOCKET_TYPE_CHANNEL_ID = "wifiSocketOutletChannel";
-
-    /**
-     * The supported thing types.
-     */
-    public final static Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = ImmutableSet.of(THING_TYPE_SOCKET_OUTLET);
-
-    // -------------- Configuration arguments ----------------
-    /**
-     * Mac address configuration argument key.
-     */
-    public final static String MAC_ADDRESS_ARG = "macAddress";
-    /**
-     * Wifi socket outlet update interval configuration argument key.
-     */
-    public final static String WIFI_SOCKET_OUTLET_UPDATE_INTERVAL_ARG = "wifiSocketOutletUpdateInterval";
-    /**
-     * Host address configuration argument key.
-     */
-    public final static String HOST_ADDRESS_ARG = "hostAddress";
-
-    // -------------- Default values ----------------
-    /**
-     * Default Wifi socket outlet refresh interval.
-     */
-    public static final long WIFI_SOCKET_OUTLET_DEFAULT_REFRESH_INTERVAL = 60;
-
-    /**
-     * Default Wifi socket outlet default UDP port.
-     */
-    public static final int WIFI_SOCKET_OUTLET_DEFAULT_UDP_PORT = 8530;
-
-    /**
-     * Discovery timeout in seconds.
-     */
-    public static final int DISCOVERY_TIMEOUT_SECONDS = 4;
-}
