--- addons/binding/org.openhab.binding.silvercrest/src/main/java/org/openhab/binding/silvercrest/wifisocketoutlet/enums/WifiSocketOutletResponseType.java
@@ -1,18 +1,0 @@
-package org.openhab.binding.silvercrest.wifisocketoutlet.enums;
-
-/**
- * This enum represents the available Wifi Socket Outlet response types.
- *
- * @author Jaime Vaz - Initial contribution
- *
- */
-public enum WifiSocketOutletResponseType {
-    /** Status changed to ON. */
-    ON,
-    /** Status changed to OFF. */
-    OFF,
-    /** ACKnowledgement. */
-    ACK,
-    /** Discovery request. */
-    DISCOVERY;
-}
