+++ addons/binding/org.openhab.binding.silvercrestwifisocket/.classpath
