+++ addons/binding/org.openhab.binding.silvercrestwifisocket/ESH-INF/binding/binding.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<binding:binding id="silvercrestwifisocket"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
+        xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
+
+    <name>Silvercrest Wifi Socket Binding</name>
+    <description>This is the binding for Silvercrest Wifi Socket sold by Lidl.</description>
+    <author>Jaime Vaz</author>
+
+</binding:binding>
