+++ addons/binding/org.openhab.binding.silvercrestwifisocket/OSGI-INF/SilvercrestWifiSocketDiscovery.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2014-2016 by the respective copyright holders.
+
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" modified="modified" name="org.openhab.binding.silvercrestwifisocket.discovery.SilvercrestWifiSocketDiscoveryService">
+   <implementation class="org.openhab.binding.silvercrestwifisocket.discovery.SilvercrestWifiSocketDiscoveryService"/>
+   <service>
+      <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
+   </service>
+</scr:component>
