+++ addons/binding/org.openhab.binding.silvercrestwifisocket/build.properties
