+++ addons/binding/org.openhab.binding.silvercrestwifisocket/src/main/java/org/openhab/binding/silvercrestwifisocket/discovery/SilvercrestWifiSocketDiscoveryService.java
@@ -1,0 +1,121 @@
+package org.openhab.binding.silvercrestwifisocket.discovery;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+
+import org.eclipse.smarthome.config.discovery.AbstractDiscoveryService;
+import org.eclipse.smarthome.config.discovery.DiscoveryResult;
+import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
+import org.eclipse.smarthome.config.discovery.DiscoveryService;
+import org.eclipse.smarthome.core.thing.Thing;
+import org.eclipse.smarthome.core.thing.ThingTypeUID;
+import org.eclipse.smarthome.core.thing.ThingUID;
+import org.openhab.binding.silvercrestwifisocket.SilvercrestWifiSocketBindingConstants;
+import org.openhab.binding.silvercrestwifisocket.handler.SilvercrestWifiSocketMediator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * This is the {@link DiscoveryService} for the Silvercrest Items.
+ *
+ * @author Jaime Vaz - Initial contribution
+ *
+ */
+public class SilvercrestWifiSocketDiscoveryService extends AbstractDiscoveryService {
+
+    private static final Logger LOG = LoggerFactory.getLogger(SilvercrestWifiSocketDiscoveryService.class);
+    private final SilvercrestWifiSocketMediator mediator;
+
+    /**
+     * Constructor of the discovery service.
+     *
+     * @throws IllegalArgumentException if the timeout < 0
+     */
+    public SilvercrestWifiSocketDiscoveryService() throws IllegalArgumentException {
+        super(SilvercrestWifiSocketBindingConstants.SUPPORTED_THING_TYPES_UIDS,
+                SilvercrestWifiSocketBindingConstants.DISCOVERY_TIMEOUT_SECONDS);
+        LOG.debug("SilvercrestWifiSocketMediator is not initialized yet will create one mediator...");
+        this.mediator = new SilvercrestWifiSocketMediator(this);
+    }
+
+    @Override
+    public Set<ThingTypeUID> getSupportedThingTypes() {
+        return SilvercrestWifiSocketBindingConstants.SUPPORTED_THING_TYPES_UIDS;
+    }
+
+    @Override
+    protected void startScan() {
+        LOG.debug("Don't need to start new scan... background scanning in progress by mediator.");
+    }
+
+    /**
+     * Method called by mediator, when receive one packet from one unknown Wifi Socket.
+     *
+     * @param macAddress the mack address from the device.
+     * @param hostAddress the host address from the device.
+     */
+    public void discoveredWifiSocket(final String macAddress, final String hostAddress) {
+        Map<String, Object> properties = new HashMap<>(2);
+        properties.put(SilvercrestWifiSocketBindingConstants.MAC_ADDRESS_ARG, macAddress);
+        properties.put(SilvercrestWifiSocketBindingConstants.HOST_ADDRESS_ARG, hostAddress);
+
+        ThingUID newThingId = this.getNewThingId(macAddress);
+        DiscoveryResult discoveryResult = DiscoveryResultBuilder.create(newThingId).withProperties(properties)
+                .withLabel("Silvercrest Wifi Socket " + newThingId.getId()).build();
+
+        LOG.debug("Discovered new thing with mac address '{}' and host address '{}'", macAddress, hostAddress);
+
+        this.thingDiscovered(discoveryResult);
+    }
+
+    /**
+     * Gets one unrepeated {@link ThingUID} based on the existing {@link ThingUID} in
+     * {@link SilvercrestWifiSocketMediator}.
+     *
+     * @param macAddress the mac address to generate the {@link ThingUID}.
+     * @return the new unexisting {@link ThingUID}.
+     */
+    private ThingUID getNewThingId(final String macAddress) {
+        ThingUID thingUID = new ThingUID(SilvercrestWifiSocketBindingConstants.THING_TYPE_WIFI_SOCKET, macAddress);
+
+        boolean thingUIDExists = this.thingUIDExistsInMediator(thingUID);
+
+        if (thingUIDExists) {
+            int index = 1;
+            while (this.thingUIDExistsInMediator(thingUID)) {
+                thingUID = new ThingUID(SilvercrestWifiSocketBindingConstants.THING_TYPE_WIFI_SOCKET,
+                        macAddress + "_" + index);
+            }
+        }
+        return thingUID;
+    }
+
+    /**
+     * Check if the {@link ThingUID} exists in {@link SilvercrestWifiSocketMediator}.
+     *
+     * @param thingUID the {@link ThingUID}.
+     * @return true if the {@link ThingUID} already exists.
+     */
+    private boolean thingUIDExistsInMediator(final ThingUID thingUID) {
+        boolean exists = false;
+        for (Thing thing : this.getMediator().getAllThingsRegistred()) {
+            if (thing.getUID().equals(thingUID)) {
+                exists = true;
+                break;
+            }
+        }
+        return exists;
+    }
+
+    // SETTERS AND GETTERS
+    /**
+     * Gets the {@link SilvercrestWifiSocketMediator} of this binding.
+     *
+     * @return {@link SilvercrestWifiSocketMediator}.
+     */
+    public SilvercrestWifiSocketMediator getMediator() {
+        return this.mediator;
+    }
+
+}
