+++ addons/binding/org.openhab.binding.silvercrestwifisocket/src/main/java/org/openhab/binding/silvercrestwifisocket/enums/SilvercrestWifiSocketResponseType.java
@@ -1,0 +1,18 @@
+package org.openhab.binding.silvercrestwifisocket.enums;
+
+/**
+ * This enum represents the available Wifi Socket response types.
+ *
+ * @author Jaime Vaz - Initial contribution
+ *
+ */
+public enum SilvercrestWifiSocketResponseType {
+    /** Status changed to ON. */
+    ON,
+    /** Status changed to OFF. */
+    OFF,
+    /** ACKnowledgement. */
+    ACK,
+    /** Discovery request. */
+    DISCOVERY;
+}
